GH-29: :building_construction: :truck: :recycle: Rename subforms and refactor implementation
