+++ app/screens/patientDetails/patientDetailsForm/PatientDemographicsSubform.stories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import PatientDemographicsSubform from './PatientDemographicsSubform';
+
+storiesOf('Screens/PatientDetails/PatientDemoGraphicsForm', module).add(
+  'default',
+  () => <PatientDemographicsSubform />
+);
