+++ app/screens/patientDetails/patientDetailsForm/PatientMedicalHistorySubform.stories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import PatientMedicalHistorySubform from './PatientMedicalHistorySubform';
+
+storiesOf('Screens/PatientDetails/PatientMedicalForm', module).add(
+  'default',
+  () => <PatientMedicalHistorySubform />
+);
