--- app/screens/patientInformation/PatientDemographicsForm.stories.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-
-import PatientDemographicsForm from './PatientDemographicsForm';
-
-storiesOf('Screens/PatientInformation/PatientDemoGraphicsForm', module).add(
-  'default',
-  () => <PatientDemographicsForm />
-);
