--- app/screens/patientInformation/PatientMedicalForm.stories.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-
-import PatientMedicalForm from './PatientMedicalForm';
-
-storiesOf('Screens/PatientInformation/PatientMedicalForm', module).add(
-  'default',
-  () => <PatientMedicalForm />
-);
